fix: error in browse panel when sort direction is initially present in query params
'this.sort', an @ViewChild, is undefined until the first render of the
component and can not be accessed during initialization.

This problem was introduced in 86e9563, which fixed a change-after-check
bug with pagination but left a use-before-ready bug with sorting. With
this commit, both are handled the same way without using @ViewChild.